requirements.txt
pydantic
semgrep
GitPython